Split services to separate modules
